+++ moor/test/database_test.dart
@@ -71,7 +71,7 @@ void main() {
     final executor = MockExecutor();
     final db = TodoDb(executor);
 
-    await db.someDao.todosForUser(1).get();
+    await db.someDao.todosForUser(user: 1).get();
 
     verify(executor.runSelect(argThat(contains('SELECT t.* FROM todos')), [1]));
   });